test: Create top-level integration tests
