+++ integration-tests/.gitignore
@@ -1,0 +1,3 @@
+node_modules/
+playwright-report/
+test-results/
