+++ integration-tests/.prettierignore
@@ -1,0 +1,3 @@
+# Dependencies
+node_modules
+package-lock.json
