+++ integration-tests/.prettierrc
@@ -1,0 +1,5 @@
+{
+  "trailingComma": "es5",
+  "semi": false,
+  "singleQuote": true
+}
