+++ integration-tests/tests/pages/HomePage.ts
@@ -1,7 +1,7 @@
 import { Locator, Page } from '@playwright/test'
 
 export class HomePage {
-  constructor(private page: Page) { }
+  constructor(private page: Page) {}
 
   async goto() {
     await this.page.goto('/')
