+++ integration-tests/tests/upload-file.spec.ts
@@ -1,13 +1,8 @@
 import { expect, test } from '@playwright/test'
-import { BackendApiMock } from './mocks/BackendApiMock'
 import { UploadAFile } from './pages/UploadAFile'
 
 test.describe('Upload a file page', () => {
   test('shows main page elements', async ({ page }) => {
-    const backendMock = new BackendApiMock(page)
-    await backendMock.abortPosthog()
-    await backendMock.mockCurrentUser()
-
     const uploadAFilePage = new UploadAFile(page)
 
     await uploadAFilePage.goto()
@@ -19,11 +14,6 @@ test.describe('Upload a file page', () => {
   })
 
   test('shows minuting controls when a file is provided', async ({ page }) => {
-    const backendMock = new BackendApiMock(page)
-    await backendMock.abortPosthog()
-    await backendMock.mockCurrentUser()
-    await backendMock.mockUserTemplates(true)
-
     const uploadAFilePage = new UploadAFile(page)
 
     await uploadAFilePage.goto()
